+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Enqueued `updateItem` mutations are now merged together to prevent the task queue from growing too much.
+
 ## [0.4.4] - 2019-10-10
 
 ### Changed